Add details on reinstallting a new proxy and switching to salt managed
With the current test it is not clear what happens if you reinstall a
proxy that was managed traditionaly to a salt one. This commit makes
this more explicity by stating that if you use the same activation key
and not remove the proxy from the list of systems, clients won't need to
be re-registered.

fixes #749
related to https://github.com/SUSE/spacewalk/issues/13257

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>