Update space-management.adoc (#2121)
Adding a blank line on line 59. It seems that the fact that there is no blank line between 'For example:' and '----', causes the formatting of the zh_CN Administration Guide PDF to become corrupted from section 9.4 onwards. Other Asian languages (JA and KO), and of course EN, were not effected by this missing blank line.