Update client-upgrades-product-migration.adoc
Request to comment-out, or remove, 3 plus signs, that are erroneously appearing in the output text.
These appear inside an IMPORTANT block. There are no lists in this block, so I believe there is no need for the list continuation + signs.
They do not operate on the formatting and instead display in the body text.