Update autoinst-distributions.adoc
Request to make a 'literal' paragraph a 'normal' paragraph. I think it was intended to be a normal paragraph originally. Lines 42 & 43. Remove the 2 preceding spaces that indent the text; "Take a note of the file path. You will need it when you declare the distribution to {productname}."
It currently formats the para in an indented green box and the {productname} entity is displaying literally instead of being resolved.