Update autoinst-profiles.adoc
Consider making a 'literal' paragraph a 'normal' paragraph. I think it was intended to be a normal paragraph originally. Line 56. Remove the 2 preceding spaces that indent the text; "For more information about the details to include here, see xref:client-configuration:autoinst-profiles.adoc#autoyast[AutoYast Profiles] or xref:client-configuration:autoinst-profiles.adoc#kickstart[Kickstart Profiles]." It currently formats the para in an indented green box and the xrefs are displaying literally instead of being resolved and functional/clickable.