Refactor SLE Micro registration warning
This will refactor the warning on the SLE Micro client page to make it
more clear that you should reboot the minion manually after e.g. the
registration via the webUI was successful. Otherwise it could take hours
until SLE Micro gets rebooted and shows up in the system list in the
webUI.

Signed-off-by: Dominik Gedon <dominik.gedon@suse.com>